Initial state for running bots on different process
* Developed a pipeline to run the bots on a different process than main.
The current problem is that since the bot is running in a different process, it is not capable of storing data, since once the process is terminated, all data is erased.
* Some possible ideas are using threads instead of multiprocessing, but there is no effective way to kill a thread.
* Another option is to send the instance as a shared object to the new process and return it, but there is no easy way to accomplish this for complex classes.
At this point, there seems to be no way to accomplish this. Further investigation into this issue is needed.