Fixed bug on the direction of new pawns
There was an error when putting new pawns on the board since it was not resetting its direction. ex. if you had a pawn with direction 1 and it was captured, and then you put it again, it was maintaining the same direction 1, instead of changing its value to -1.
Now, each time that you put a new pawn on the board it sets its direction to -1 independently of where you put it.